+++ lightbulb/errors.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Copyright © Thomm.o 2020
+# Copyright © Thomm.o 2021
 #
 # This file is part of Lightbulb.
 #
@@ -101,9 +101,11 @@ class NotEnoughArguments(CommandError):
     Exception raised when a command is run without a sufficient number of arguments.
     """
 
-    def __init__(self, command: commands.Command) -> None:
+    def __init__(self, command: commands.Command, missing_args: typing.List[str]) -> None:
         self.command: commands.Command = command
         """The command string that was attempted to be invoked."""
+        self.missing_args: typing.List[str] = missing_args
+        """The required arguments that are missing."""
 
 
 class TooManyArguments(CommandError):